Add fetches and compile classes
