+++ src/Traveler.js
@@ -1,10 +1,24 @@
+import { travelerData } from "./data/traveler-data";
+
 class Traveler {
-  constructor(travelerData, travelManager) {
+  constructor(travelerData) {
     this.id = travelerData.id;
     this.name = travelerData.name;
+    this.trips = [];
     this.travelerType = travelerData.travelerType;
-    this.travelManager = travelManager;
   }
-}
 
+  findTravelerTrips(tripData) {
+    this.trips = tripData.filter((trip)=> trip.userID === this.id);
+  }
+
+  getTravelerById(id) {
+    return this.travelers.find((traveler) => traveler.id === id);
+  }
+
+  getTravelersByType(type) {
+    return this.travelers.filter((traveler) => traveler.travelerType === type);
+  }
+
+}
 export default Traveler;
