+++ src/apiCalls.js
@@ -7,46 +7,42 @@ const fetchAPI = (url) =>
   .catch((error) => console.log(error))
 
 
-const fetchAllData = () => {
+ const fetchAllData = () => {
   return Promise.all([
-  fetchAPI(`http://localhost:3001/api/v1/travelers/`),
+  fetchAPI(`http://localhost:3001/api/v1/travelers/1`),
   fetchAPI('http://localhost:3001/api/v1/trips'),
   fetchAPI('http://localhost:3001/api/v1/destinations')
   ]);
 };
   
-const fetchSingleTraveler = (id) => {
-  return fetch(`http://localhost:3001/api/v1/travelers/${id}`)
-  .then(res => {
-    if (!res.ok) {
-      throw new Error("Missing or Incorrect Username/Password");
-    } else {
-      return res.json()
-    }
-  })
-  .catch(() => {
-    return false
-  });
-}
+// const fetchSingleTraveler = (id) => {
+//   return fetch(`http://localhost:3001/api/v1/travelers/${id}`) put this back on 12!
+//   .then(res => {
+//     if (!res.ok) {
+//       throw new Error("Missing or Incorrect Username/Password");
+//     } else {
+//       return res.json()
+//     }
+//   })
+//   .catch(() => {
+//     return false
+//   });
+// }
 
-const posttripsData = (data) => {
-  return fetch('http://localhost:3001/api/v1/trips', {
-    method: 'POST',
-    body: JSON.stringify(data),
-    headers: {
-      'Content-Type': 'application/json'
-    }
-  })
-  .then(res => {
-    if (!res.ok) {
-      throw new Error('There seems to be an error!')
-    }
-    return res.json()
-  }).catch(error => console.log(error))
-};
+// const posttripsData = (data) => {
+//   return fetch('http://localhost:3001/api/v1/trips', {
+//     method: 'POST',
+//     body: JSON.stringify(data),
+//     headers: {
+//       'Content-Type': 'application/json'
+//     }
+//   })
+//   .then(res => {
+//     if (!res.ok) {
+//       throw new Error('There seems to be an error!')
+//     }
+//     return res.json()
+//   }).catch(error => console.log(error))
+// };
 
-export default { 
-fetchAllData,
-fetchSingleTraveler, 
-posttripsData
-};
+export { fetchAllData };
